+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ponder/react
 
+## 0.14.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.14.7
+
+## 0.14.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.14.6
+
 ## 0.14.5
 
 ### Patch Changes